% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ahrqComorbid}
\alias{ahrqComorbid}
\title{AHRQ comorbidities}
\format{list of character vectors}
\source{
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp}
}
\description{
This mapping of comorbidities to ICD-9 codes is dervied directly
  from SAS code provided by AHRQ, and translated into this R data structure.
}
\keyword{datasets}

