% Generated by roxygen2 (4.0.1): do not edit by hand
\name{allIsNumeric}
\alias{allIsNumeric}
\title{check whether character vector represents all numeric values}
\usage{
allIsNumeric(x, extras = c(".", "NA"))
}
\arguments{
\item{x}{is a character vector to be tested}

\item{extras}{is a vector of character strings which are counted as NA
values, defaults to '.' and 'NA'}
}
\value{
logical
}
\description{
check whether all the items of input vector are numeric without
  throwing warning derived from Hmsic package
}
\keyword{internal}

