% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asNumericNoWarn}
\alias{areIntegers}
\alias{asIntegerNoWarn}
\alias{asNumericNoWarn}
\title{convert factor or vector to numeric without warnings}
\usage{
asNumericNoWarn(x)

asIntegerNoWarn(x)

areIntegers(x)
}
\arguments{
\item{x}{is a vector, probably of numbers of characters}
}
\value{
numeric vector, may have NA values
}
\description{
correctly converts factors to vectors, and then converts to
  numeric or integer, which may silently introduce NAs. Invisible rounding
  errors can be a problem going from numeric to integer, so consider adding
  tolerance to this conversion. \code{asIntegerNoWarn} silently \code{\link{floor}}s.
}
\keyword{internal}

