% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getInvalidDecimalIcd9}
\alias{getInvalidDecimalIcd9}
\alias{getInvalidShortICD9}
\alias{getInvalidShortIcd9}
\title{invalid subset of decimal or short ICD-9 codes}
\usage{
getInvalidDecimalIcd9(icd9Decimal)

getInvalidShortICD9(icd9Short)
}
\arguments{
\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than five characters is given
in a code, then the digits are greedily assigned to hundreds, then tens,
then units, before the decimal parts. E.g. "10" becomes "010", not "0010"}
}
\description{
given vector of short or decimal ICD-9 codes in
  \code{icd9Decimal} or \code{icd9Short}, return (in the same format) those
  codes which are invalid. Useful for generating error messages with the
  faulty codes if validation fails.
}
\keyword{manip}

