% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%i9mj\%}
\alias{\%i9mj\%}
\title{create range of icd9 major parts}
\usage{
start \%i9mj\% end
}
\arguments{
\item{start,end}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\value{
character vector with range inclusive of start and end
}
\description{
accepts V, E or numeric codes. Does not validate codes beyond
  ensuring that the start and end of the range are of the same type.
}
\seealso{
Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{icd9ExpandRangeDecimal}};
  \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9ChildrenDecimal}};
  \code{\link{icd9Children}},
  \code{\link{icd9ChildrenShort}};
  \code{\link{icd9CondenseShort}};
  \code{\link{icd9CondenseToExplainShort}};
  \code{\link{icd9ExpandMinor}},
  \code{\link{icd9ExpandMinorE}},
  \code{\link{icd9ExpandMinorNV}};
  \code{\link{icd9PrecedingMinors}},
  \code{\link{icd9SubsequentMinors}}
}

