% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%nin\%}
\alias{\%nin\%}
\title{inverse of \%in\%}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{is the vector of values to be matched}

\item{table}{is actually a vector, to be matched against}
}
\value{
logical vector of length of x
}
\description{
borrowed from Hmisc. See %in%
}
\keyword{internal}

