% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9AddLeadingZeroesDecimal}
\alias{icd9AddLeadingZeroesDecimal}
\title{pad decimal ICD-9 codes with leading zeroes}
\usage{
icd9AddLeadingZeroesDecimal(icd9Decimal, invalidAction = c("ignore", "silent",
  "warn", "stop"))
}
\arguments{
\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}
}
\value{
character vector of icd9 codes
}
\description{
pad decimal ICD-9 codes with leading zeroes
}
\keyword{internal}
\keyword{manip}

