% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9AddLeadingZeroesMajor}
\alias{icd9AddLeadingZeroesMajor}
\title{zero-pad major part of ICD9 code}
\usage{
icd9AddLeadingZeroesMajor(major)
}
\arguments{
\item{major}{character vector of 'major' part of ICD-9 codes, i.e. that part
which falls before the decimal point, in decimal notation. (In 5 digit
notation, the 'major' part is be the first three characters (with leading
zeroes), and includes V or E prefix. xyz}
}
\value{
character vector
}
\description{
three digit codes are returned unchanged, one and two digit
  codes are preceded by 00 or 0. V or E codes are trimmed for whitespace and
  returned without prefxing. NA values are generated where the code is
  non-numeric, and doesn't match a valid V or E code format.
}
\keyword{internal}
\keyword{manip}

