% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9AddLeadingZeroesShort}
\alias{icd9AddLeadingZeroesShort}
\title{add leading zeroes to short-form ICD-9 code}
\usage{
icd9AddLeadingZeroesShort(icd9)
}
\arguments{
\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\description{
Use with care. non-decimal ICD-9 codes with length<5 are often
  ambiguous. E.g. 100 could be 1.00 10.0 or 100
}
\keyword{internal}
\keyword{manip}

