% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9ChapterShort}
\alias{icd9ChapterShort}
\title{get ICD-9 Chapters from vector of ICD-9 codes}
\usage{
icd9ChapterShort(icd9Short, isShort, invalidAction = icd9InvalidActions)
}
\arguments{
\item{icd9-any}{}

\item{isShort}{}

\item{invalid}{}
}
\description{
work-in-progress
}
\keyword{internal}

