% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9ExpandBaseCodeShort}
\alias{icd9ExpandBaseCode}
\alias{icd9ExpandBaseCodeShort}
\title{expand 5 character form 'short' ICD9 to all possible sub codes}
\usage{
icd9ExpandBaseCodeShort(icd9Short, validate = F)

icd9ExpandBaseCode(icd9, short)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than five characters is given
in a code, then the digits are greedily assigned to hundreds, then tens,
then units, before the decimal parts. E.g. "10" becomes "010", not "0010"}

\item{validate}{logical whether to validate the input ICD-9 codes for
correctness. There may be additional simple checks, e.g. whether the code
is of the right type, regardless of this argument.}

\item{short}{single logical value which determines whether the ICD-9 code provided is in short (TRUE) or decimal (FALSE) form.}

\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{short}{single logical value which determines whether the ICD-9 code provided is in short (TRUE) or decimal (FALSE) form.}
}
\description{
this is so the raw info from SAS code provided by AHRQ can be
  interpreted without manually reformatting.
}
\keyword{internal}

