% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9ExpandRangeDecimal}
\alias{\%i9d\%}
\alias{icd9ExpandRangeDecimal}
\title{expand range of ICD-9 decimal codes to all possible intermediate and sub-codes}
\usage{
icd9ExpandRangeDecimal(start, end)

start \%i9d\% end
}
\description{
As with \code{link{icd9ExpandRangeShort}} great care is taken not to include codes which have children not in the range. E.g. "100.9" to "101.1" would _not_ include code "101".
}
\seealso{
Other ICD-9: \code{\link{\%i9mj\%}}

Other jack: \code{\link{\%i9mj\%}}
}
\keyword{internal}
\keyword{manip}

