% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{icd9InvalidActions}
\alias{icd9InvalidActions}
\title{invalid actions, default is first item}
\format{\preformatted{ chr [1:4] "ignore" "silent" "warn" "stop"
}}
\usage{
icd9InvalidActions
}
\description{
Some functions accept \code{invalidAction} argument, but require non-default validation, but the vast majority will honor the default of 'ignore.' Ignore may lead to downstream errors, since bad data may arrive in internal functions, however, this is the fastest option. A typical use-case of this package would involve validation and cleaning steps (with validation), followed by repeated analysis (as fast as possible, with minimal validation).
}
\keyword{character}
\keyword{sysdata}

