% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9ParseAndSaveMappings}
\alias{icd9ParseAndSaveMappings}
\title{parse all known mappings and save to development tree}
\usage{
icd9ParseAndSaveMappings(path = "~/icd9/data")
}
\arguments{
\item{path}{directory to save in, default is \code{~/icd9/data}}
}
\description{
parse all known mappings and save to development tree
}
\keyword{internal}

