% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9Tell}
\alias{icd9Tell}
\title{describe ICD-9 codes, guess whether short or long}
\usage{
icd9Tell(icd9)
}
\description{
partly implemented. Goal is to guess whether codes are short or
  decimal form, then to call icd9Explain with the condense argument.
  Currently condense works, but not with the icd9 lookup table currently in
  use. Not exporting this function until it works as intended.
}
\keyword{internal}

