% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9WarnStopMessage}
\alias{icd9WarnStopMessage}
\title{warn or stop with message based on switch}
\usage{
icd9WarnStopMessage(..., invalidAction = c("stop", "warn", "silent"))
}
\arguments{
\item{...}{message components passed on to \code{warning} or \code{stop}}

\item{invalid}{}
}
\description{
warn or stop with message based on switch
}
\keyword{internal}

