% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logicalToBinary}
\alias{logicalToBinary}
\title{encode TRUE as 1, and FALSE as 0}
\usage{
logicalToBinary(dframe)
}
\arguments{
\item{dframe}{dataframe which may contain logical fields}
}
\value{
dframe without logical fields
}
\description{
when saving data as text files for distribution, printing large
  amounts of text containing TRUE and FALSE is inefficient. Convert to binary
  takes more R memory, but allows more compact output TODO: test
}
\keyword{internal}
\keyword{manip}

