% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{quanDeyoComorbid}
\alias{quanDeyoComorbid}
\title{Quan adaptation of Deyo/Charlson comorbidities}
\format{list of character vectors, each named by co-morbidity}
\description{
Derived programmatically from the SAS code used in the original
  publication. According to the referenced study, this provides the best
  presdictor of in-patient to <30d mortality.
}
\references{
Quan, Hude, Vijaya Sundararajan, Patricia Halfon, Andrew Fong,
  Bernard Burnand, Jean-Christophe Luthi, L. Duncan Saunders, Cynthia A.
  Beck, Thomas E. Feasby, and William A. Ghali. "Coding Algorithms for
  Defining Comorbidities in ICD-9-CM and ICD-10 Administrative Data." Medical
  Care 43, no. 11 (November 1, 2005): 1130-39.
  \url{http://www.ncbi.nlm.nih.gov/pubmed/16224307}
}
\keyword{datasets}

