% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveSourceTreeData}
\alias{saveSourceTreeData}
\title{save data in source tree}
\usage{
saveSourceTreeData(varName, path = "~/icd9/data")
}
\arguments{
\item{varName}{is the variable name and the part of the filename which will
be saved, followed by ".RData"}

\item{path}{is a path name to destination folder for the data: no trailing slash.}
}
\description{
attempt to write the date from the source file to RData in the
  package source tree.
}

