% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strip}
\alias{strip}
\title{strip all whitespace}
\usage{
strip(x)
}
\arguments{
\item{x}{is a character vector to strip}
}
\value{
character vector
}
\description{
could do this with regular expression, but slow, and this
  function is called frequently. My only use case works with removal of all
  whitespace.
}
\keyword{internal}

