% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trim}
\alias{trim}
\title{strip whitespace from ends of each string in given character vector}
\usage{
trim(x)
}
\arguments{
\item{x}{is a character vector to trim}
}
\value{
character vector
}
\description{
slower than \code{strip}.
}
\keyword{internal}

