% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%i9mj\%}
\alias{\%i9mj\%}
\title{find range of ICD-9 codes between two E-codes}
\usage{
start \%i9mj\% end
}
\arguments{
\item{start,end}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\value{
character vector with range inclusive of start and end
}
\description{
E codes only have a single possible numeral after the decimal
  place. This makes processing very different from V and pure numeric codes.
}
\seealso{
Other ICD-9: \code{\link{\%i9d\%}},
  \code{\link{icd9ExpandRangeDecimal}}

Other jack: \code{\link{\%i9d\%}},
  \code{\link{icd9ExpandRangeDecimal}}
}

