% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{icd9CmDesc}
\alias{icd9CmDesc}
\title{ICD9-CM diagnosis code lookup}
\format{data frame}
\source{
\url{http://www.cdc.gov/nchs/icd/icd9cm.htm}
}
\description{
short-form ICD-9 codes with short and long descriptions
}
\keyword{datasets}

