% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icd9DropZeroFromDecimal}
\alias{icd9DropZeroFromDecimal}
\title{drop zero padding from decimal ICD-9 code.}
\usage{
icd9DropZeroFromDecimal(icd9)
}
\arguments{
\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\value{
character vector of ICD-9 codes with extra zeroes dropped from major
  part
}
\description{
decimal form ICD-9 codes are not ambiguous if the leading zeroes
  are dropped. Some short-form ICD-9 codes would not be ambiguous, e.g. "1"
  but many would be problematic, so no function is provided to do this for
  short-form codes.
}
\keyword{internal}
\keyword{manip}

