% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parseIcd9Chapters}
\alias{parseIcd9Chapters}
\title{parse list of top-level ICD-9 chapters from canonical data from CDC.}
\usage{
parseIcd9Chapters()
}
\value{
named list, with name of chapter being the item name, and value being the three-digit code, or Vxx or Exxx code.
}
\description{
There is no easily machine-readable list of the three digit (I call the 'major') ICD-9 code chapters. This code downloads a pretty RTF file and extracts these codes with their names. WORK IN PROGRESS!
}
\keyword{internal}

